requirements.txt
streamlit
pandas
pysqlite3-binary
mysql-connector
matplotlib
google-generativeai
streamlit-option-menu
python-dotenv
